feat(graph): Add modified Dijkstra's for Number of Ways to Arrive
This commit introduces a solution to the "Number of Ways to Arrive at Destination" problem (LeetCode #1976).

The solution uses a modified version of Dijkstra's algorithm to not only find the shortest path but also to count the number of unique paths that have this shortest length.

In addition to the standard `distance` array, a `ways` array is maintained.
- When a new shorter path to a node is found, its `ways` count is updated from its predecessor.
- When an alternative path of the same shortest length is found, the `ways` from the predecessor are added to the node's existing `ways` count.

This correctly propagates the number of shortest paths through the graph.